anvil: switch from gradle script to gradle plugin for DSL generation
Before this commit, the entire code generation logic was all crammed
into one file. While this is a fine solution when there is a single
artefact and project utilising the script, as soon as multiple
projects wish to start using the code, managing the system becomes
a lot hard.

Make the script into a plugin which can be applied to local projects
to generate code.

This commit lays the groundwork for introduction of code generation
for the support libraries (i.e. issues https://github.com/zserge/anvil/issues/34
and https://github.com/zserge/anvil/issues/35).