anvil-gen: advance dependency management and fix code generation for strange classes
The first thing this commit does is increase the flexibility of
dependencies - it means that libraries can have multiple jars of the
same project that they depend on. AppCompat is updated to make use of
this by referring to the internal library jar of support v4.

Next, this commit moves from using Java style get{Blah} calls when
working with classes to {blah} parameter accesses during code
generation. Now this may seem like a style change but for some black
magic reason, the former tries to instantiate the class and the latter
doesn't!!! This is even though both should just call the get{Blah}
function. I'm not sure why this happens - if someone else who is more
knowledgable on Groovy sees this please let me know!

This commit also allows enforces that all methods considered have public
method parameters so a compile error would not be caused when invoking
the method outside the package - this removes the need for manual quirks
for these sorts of classes.

Finally, this commit regenerates the appcompat-v7 library DSL adding a
bunch of new methods which was possible using all of the above :)