anvilgen: move from Groovy to Kotlin
While Groovy is considered the de-facto language for Gradle plugin
development, writing a Kotlin implementation revealed several subtle
bugs which where being hidden by Groovy and overall seems to make the
code clearer in most places (with the exception of the ugly closure
implementations which are unfortunately required to interface with the
gradle API).

Switch everything over to Kotlin in a manner where no functional changes
are made. This can be verified by regenerating all the DSLs and checking
that nothing is changed in that case.